refactors to use more consistent directory structure
